@@ -1,37 +1,69 @@
 {
   "name": "actual-sync",
-  "version": "1.0.1",
+  "version": "25.2.1",
   "license": "MIT",
   "description": "actual syncing server",
-  "main": "index.js",
+  "type": "module",
   "scripts": {
     "start": "node app",
-    "lint": "eslint .",
+    "lint": "eslint . --max-warnings 0",
+    "lint:fix": "eslint . --fix",
     "build": "tsc",
+    "test": "NODE_ENV=test NODE_OPTIONS='--experimental-vm-modules --trace-warnings' jest  --coverage",
+    "db:migrate": "NODE_ENV=development node src/run-migrations.js up",
+    "db:downgrade": "NODE_ENV=development node src/run-migrations.js down",
+    "db:test-migrate": "NODE_ENV=test node src/run-migrations.js up",
+    "db:test-downgrade": "NODE_ENV=test node src/run-migrations.js down",
     "types": "tsc --noEmit --incremental",
-    "verify": "yarn -s lint && yarn types"
+    "verify": "yarn lint && yarn types",
+    "reset-password": "node src/scripts/reset-password.js",
+    "enable-openid": "node src/scripts/enable-openid.js",
+    "disable-openid": "node src/scripts/disable-openid.js",
+    "health-check": "node src/scripts/health-check.js"
   },
   "dependencies": {
-    "@actual-app/api": "4.1.0",
-    "@actual-app/web": "4.1.0",
-    "bcrypt": "^5.0.1",
-    "better-sqlite3": "^7.5.0",
-    "body-parser": "^1.18.3",
+    "@actual-app/crdt": "2.1.0",
+    "@actual-app/web": "25.2.1",
+    "bcrypt": "^5.1.1",
+    "better-sqlite3": "^11.7.0",
+    "body-parser": "^1.20.3",
     "cors": "^2.8.5",
-    "express": "4.17",
-    "express-actuator": "^1.8.1",
+    "date-fns": "^2.30.0",
+    "debug": "^4.3.4",
+    "express": "4.20.0",
+    "express-actuator": "1.8.4",
+    "express-rate-limit": "^6.7.0",
     "express-response-size": "^0.0.3",
-    "node-fetch": "^2.2.0",
-    "uuid": "^3.3.2"
+    "express-winston": "^4.2.0",
+    "jws": "^4.0.0",
+    "migrate": "^2.0.1",
+    "nordigen-node": "^1.4.0",
+    "openid-client": "^5.4.2",
+    "uuid": "^9.0.0",
+    "winston": "^3.14.2"
   },
   "devDependencies": {
-    "@types/better-sqlite3": "^7.5.0",
-    "@types/node": "^17.0.31",
-    "@typescript-eslint/eslint-plugin": "^5.23.0",
-    "@typescript-eslint/parser": "^5.23.0",
-    "eslint": "^8.15.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "prettier": "^2.6.2",
-    "typescript": "^4.6.4"
-  }
+    "@babel/preset-typescript": "^7.20.2",
+    "@types/bcrypt": "^5.0.2",
+    "@types/better-sqlite3": "^7.6.12",
+    "@types/cors": "^2.8.13",
+    "@types/express": "^4.17.17",
+    "@types/express-actuator": "^1.8.0",
+    "@types/jest": "^29.2.3",
+    "@types/node": "^17.0.45",
+    "@types/supertest": "^2.0.12",
+    "@types/uuid": "^9.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.51.0",
+    "@typescript-eslint/parser": "^5.51.0",
+    "eslint": "^8.33.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^29.3.1",
+    "prettier": "^2.8.3",
+    "supertest": "^6.3.1",
+    "typescript": "^4.9.5"
+  },
+  "engines": {
+    "node": ">=18.0.0"
+  },
+  "packageManager": "yarn@4.3.1"
 }
